LayoutTests/imported/w3c/web-platform-tests/css/css-properties-values-api/animation/custom-property-transition-transform-function-to-list-expected.txt


FAIL A custom property cannot yield a CSS Transition from <transform-function> to <transform-list> assert_equals: A single animation is running expected 1 but got 0
